// src/app/models/chamado.ts
export interface Chamado {
    id?:                any;
    dataAbertura?:   string;
    dataFechamento?: string;
    prioridade:      string;
    status:          string;
    titulo:          string;
    observacoes:     string;
    estabelecimento:            any;
    motoboy:            any;
    nomeEstabelecimento:     string;
    nomeMotoboy:     string;
 }